move constants to appropriate file
