+++ src/applications/gi/constants.js
@@ -1,2 +1,6 @@
 // WAIT_INTERVAL is in milliseconds.
 export const WAIT_INTERVAL = 333;
+
+export const ARIA_LABELS = Object.freeze({
+  preferredProviderLearnMore: 'Learn more about VET TEC Preferred Providers',
+});
