--- src/applications/gi/utils/constants.js
@@ -1,3 +1,0 @@
-export const ARIA_LABELS = Object.freeze({
-  PREFERRED_PROVIDER_LEARN_MORE: 'Learn more about VET TEC Preferred Providers',
-});
